Add RIS files from first search
From the definition of keywords, combinations were created between the two research axes, resulting in 10 combinations. The Boolean operator used for these combinations was “AND,” which returns results including each of the keywords from both axes in the search fields. In the search procedure for the selected databases, Scopus and Web of Science, additional criteria were applied, such as a temporal limit of 10 years (2014–2024), publication type (Article only), and article language (English only).